api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-dotenv
alembic
psycopg2-binary
psycopg
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt
pydantic-settings
email-validator
sqlalchemy
openai
mcp
google-generativeai
requests
httpx
psycopg[binary]